refactor(web): #362 realign to agreed 3 zones (Sources / Composition / Diagram)
The drawer had drifted into a Sources→Discovery→Mapping→Resolved stepper — i.e.
the old five tabs reskinned, the opposite of the agreed 3-zone collapse
(composition-store doc § "Target (3 tabs)"). Realigned:

- Shell: the drawer is a two-way zone switch **Sources | Composition** (+ gear).
  No "Discovery" — it's renamed Composition. Within Composition, Mapping and
  Resolved are **subviews** (Entities | Mapping | Resolved), not top-level tabs.
- Route `discovery/` → renamed to `composition/` (the Entities/curation surface);
  `discovery/+page.ts` 308-redirects old links to `/composition`. mapping/ and
  resolved/ stay as Composition subviews.
- Sync + scope moved to **Sources** (the input zone owns ingestion): per-source
  Sync now, Sync all, Rebuild, and the faceted Scope panel now live on Sources,
  next to the config they apply to — no cross-surface seam. They were removed
  from the Composition page.
- Composition page is now purely curation: resolved-entity grid + per-entity
  detail (policy) + observations (+ Mapping/Resolved subviews).

Design doc (topology-ui-ia.md) rewritten to match (zones, no stepper, sync/scope
in Sources, Composition entity-centric). Deferred: folding metrics-binding fully
into per-entity detail (Mapping is a subview for now); left-rail vs drawer for
Sources; manual-editor relocation.
